pyproject.toml
[tool.poetry]
name = "axians-netbox-plugin-pdu"
version = "0.0.2"
description = "A plugin for NetBox to easily display PDU information."
authors = ["Alexander Gittings <alexander.gittings@axians.co.uk>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/minitriga/axians-netbox-plugin-pdu"
repository = "https://github.com/minitriga/axians-netbox-plugin-pdu"
keywords = ["netbox", "network", "pdu", "django"]
include = [
    "LICENSE",
    "README.md"
]
packages = [
    { include = "axians_netbox_pdu" },
]

[tool.poetry.dependencies]
python = "^3.6 || ^3.7 || ^3.8"
invoke = "^1.4.1"
easysnmp = "^0.2.5"
rq-scheduler = "^0.10.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
yamllint = "^1.23.0"
bandit = "^1.6.2"
pylint = "^2.5.2"
pylint-django = "^2.0.15"
pydocstyle = "^5.0.2"
easysnmp = "^0.2.5"
rq-scheduler = "^0.10.0"

[tool.black]
line-length = 120
target-version = ['py36']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | settings.py     # This is where you define files that should not be stylized by black
                     # the root of the project
)
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
